Move util functions to util class
